+++ README.md
@@ -21,7 +21,7 @@
 
 `예외`
 
-- [x] 존재하지 않는 회원일 경우 - **400**
+- [x] 존재하지 않는 회원일 경우 - **404**
 - [x] 인가되지 않은 요청일 경우 - **401**
 
 ### 회원가입 기능
@@ -47,8 +47,8 @@
 `예외`
 
 - [x] 이메일, 비밀번호가 입력되지 않은 경우 - **400**
-- [x] 이메일이 존재하지 않는 경우 - **400**
-- [x] 비밀번호가 일치하지 않는 경우 - **400**
+- [x] 이메일이 존재하지 않는 경우 - **401**
+- [x] 비밀번호가 일치하지 않는 경우 - **401**
 
 ### 회원 수정 기능
 
@@ -65,7 +65,7 @@
 
 `예외`
 
-- [x] 존재하지 않는 회원일 경우 - **400**
+- [x] 존재하지 않는 회원일 경우 - **404**
 
 ### 회원 삭제 기능
 
@@ -74,7 +74,7 @@
 
 `예외`
 
-- [x] 비밀번호가 일치하지 않는 경우 - **400**
+- [x] 비밀번호가 일치하지 않는 경우 - **401**
 
 ## 사용자 시나리오
